requirements.txt
Django==2.2.7
django-simple-cookie-consent==0.1.1
docxtpl==0.6.3
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-docx==0.8.7
pytz==2019.3
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
